requirements.txt
streamlit==1.28.1
pandas==2.1.1
pymongo==4.5.0
plotly==5.17.0
requests==2.31.0
scikit-learn==1.3.0
numpy==1.24.3
python-dotenv==1.0.0
fdpdf
